backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Reuse existing dependencies
pydantic>=2.11.7
loguru==0.7.2

# Additional dependencies
requests==2.31.0
email-validator==2.1.0

# Supabase
supabase==2.27.3
